refactor(driver): consolidate sidecar eigensolve drivers (#144)
Extend reference/driver/eigensolve_from_sidecar.py to accept
--problem cube-cavity|sphere-pec (default cube-cavity for backwards
compatibility) in addition to the existing --backend tfjava|onnx,
collapsing the (problem x backend) fragmentation called out in #144.

The sphere-PEC code path is ported verbatim from
eigensolve_sphere_pec_sidecar.py: shift-and-invert ARPACK at sigma=0,
d-zero rank spurious-mode classifier (PR #126), spurious filter, and
the optional baseline cross-check at --rtol.

The three legacy entry points are reduced to thin shims that inject
the right --problem / --backend flags and delegate here:

  * eigensolve_from_tfjava.py     -> --backend tfjava
  * eigensolve_from_onnx.py       -> --backend onnx
  * eigensolve_sphere_pec_sidecar.py -> --problem sphere-pec (default
    --backend tfjava, honoring an explicit override)

Each shim's docstring now carries a deprecation note pointing at the
consolidated entry point. CI workflows
(onnx-cube-cavity.yml / tfjava-cube-cavity.yml) keep working
unchanged because the shims absorb the new flag.

README updated with the consolidated entry-point usage and a shim
deprecation table.

Smoke tests (all local):
  * cube-cavity onnx + tfjava: byte-for-byte parity with previous
    eigensolve_from_sidecar.py output
  * sphere-pec via tfjava and onnx backends: physical eigenvalues
    agree with reference/fixtures/sphere_pec/baseline.json to < 1e-5
    relative on the lowest 5 modes (full 1e-5 Epic #88 cross-IR gate
    still passes). Inter-run drift is ~1e-11 / 1e-10 relative, which
    is the same ARPACK indeterminacy seen between two consecutive
    runs of the pre-refactor driver -- no regression introduced.
  * cargo test -p geode-validation: 16 passed, 5 ignored, 0 failed.

Closes #144
