+++ reference/driver/eigensolve_sphere_pec_sidecar.py
@@ -1,330 +1,29 @@
-"""SciPy eigensolve driver for sphere-PEC Nédélec sidecars (Epic #88 / issue #134).
+"""Thin wrapper — delegates to eigensolve_from_sidecar.py --problem sphere-pec.
 
-Consumes the JSON sidecar produced by the TF-Java sphere-PEC assembly
-(reference/tf_java/sphere_pec/SpherePecMain) and runs the full
-shift-and-invert ARPACK eigensolve, spurious-mode classification via the
-algebraic d⁰ rank (Issue #124), and spurious filtering — then emits a
-schema-v1 fixture with the lowest physical eigenvalues.
+Deprecated: use ``eigensolve_from_sidecar.py --problem sphere-pec
+--backend <tfjava|onnx>`` directly. Preserved for backward compatibility
+so CI workflows and external callers referencing this path continue to
+work unchanged. All logic lives in the consolidated
+``eigensolve_from_sidecar.py`` (issue #144).
 
-The corresponding cube-cavity driver is eigensolve_from_sidecar.py; that
-script is not reused here because the sphere-PEC pipeline requires:
-
-  1. The full (K_int, M_int) dense matrices (embedded in the sidecar).
-  2. Shift-and-invert at sigma=0 (ARPACK eigsh, not dense eigh) to recover
-     the large spurious null-space before the physical modes.
-  3. The d⁰-rank spurious classifier (reference/numpy/sphere_pec.py::
-     spurious_dim_from_derham), which requires the interior-node mask and
-     edge table — NOT available from the cube-cavity sidecar format.
-     For the TF-Java sphere-PEC sidecar we read n_spurious_predicted
-     (= interior node count, stored by the Java main as `spurious_dim`)
-     and use it as the shift point; the d⁰-rank algebraic count is then
-     re-computed in the Rust validation harness from the Burn-side mesh.
-
-Usage
-=====
-    python3 reference/driver/eigensolve_sphere_pec_sidecar.py \\
-        path/to/reduced_kM_sphere_pec.json \\
-        [--k 5] [--n-request-extra 8] [--out eigenresult_sphere_pec.json]
+The shim defaults ``--backend tfjava`` to match the historical behavior
+of this script (sphere-PEC was TF-Java-only when it was first added in
+PR #137). Callers that want a different backend may pass
+``--backend onnx`` (or any other accepted value) explicitly — the shim
+forwards extra arguments verbatim.
 """
 
-from __future__ import annotations
-
-import argparse
-import json
 import sys
+import runpy
 from pathlib import Path
 
-import numpy as np
-import scipy.sparse as sp
-import scipy.sparse.linalg as spla
+extra = sys.argv[1:]
+# Inject sensible defaults that the legacy script provided implicitly. We
+# only inject them when the caller has NOT already supplied them, so
+# explicit overrides from CI / external callers always win.
+forwarded = ["--problem", "sphere-pec"]
+if "--backend" not in extra:
+    forwarded += ["--backend", "tfjava"]
 
-
-def _load_field(sidecar: dict, section: str, key: str) -> np.ndarray:
-    field = sidecar[section][key]
-    arr = np.asarray(field["data"], dtype=np.float64).ravel()
-    shape = field["shape"]
-    expected = int(np.prod(shape))
-    if arr.size != expected:
-        raise ValueError(
-            f"Sidecar {section}/{key}: data length {arr.size} "
-            f"does not match shape {shape} (= {expected})"
-        )
-    return arr.reshape(shape)
-
-
-def main():
-    parser = argparse.ArgumentParser(
-        description=(
-            "SciPy shift-and-invert eigensolve over a TF-Java sphere-PEC sidecar. "
-            "Applies the spurious-mode filter and emits a schema-v1 fixture with "
-            "the lowest physical eigenvalues."
-        )
-    )
-    parser.add_argument(
-        "sidecar",
-        help="Path to the reduced_kM_sphere_pec.json sidecar from the TF-Java assembly.",
-    )
-    parser.add_argument(
-        "--k",
-        type=int,
-        default=5,
-        help="Number of physical eigenvalues to return (default 5).",
-    )
-    parser.add_argument(
-        "--n-request-extra",
-        type=int,
-        default=8,
-        help=(
-            "Extra modes to request beyond the predicted spurious count "
-            "(default 8, same as reference/numpy/sphere_pec.py::run_sphere_pec)."
-        ),
-    )
-    parser.add_argument(
-        "--n-spurious",
-        type=int,
-        default=None,
-        help=(
-            "Override spurious-mode count (use if the sidecar does not embed it). "
-            "If omitted, read from sidecar['inputs']['n_int']['data'][0] which "
-            "equals the interior edge count — NOT the interior node count needed "
-            "for the spurious filter. For the TF-Java sidecar, the Java driver "
-            "stores 'spurious_dim' as a separate output if available; otherwise "
-            "use this override."
-        ),
-    )
-    parser.add_argument(
-        "--baseline",
-        default=None,
-        help=(
-            "Path to reference/fixtures/sphere_pec/baseline.json; if provided, "
-            "print a cross-backend comparison table."
-        ),
-    )
-    parser.add_argument(
-        "--rtol",
-        type=float,
-        default=2e-5,
-        help=(
-            "Relative tolerance for the cross-backend eigenvalue gate (default 2e-5). "
-            "TF-Java's scatterNd accumulation order differs from NumPy's COO->CSR path, "
-            "producing ~1.2e-5 relative error on the lowest eigenvalue; 2e-5 gives "
-            "adequate margin. Pass 1e-5 to enforce the strict Epic #88 cross-IR floor."
-        ),
-    )
-    parser.add_argument("--out", default="eigenresult_sphere_pec.json")
-    args = parser.parse_args()
-
-    sidecar_path = Path(args.sidecar)
-    if not sidecar_path.exists():
-        print(f"Sidecar not found: {sidecar_path}", file=sys.stderr)
-        sys.exit(2)
-
-    with open(sidecar_path) as f:
-        sidecar = json.load(f)
-
-    # --- Load K_int, M_int ---
-    k_int_flat = _load_field(sidecar, "outputs", "k_int")
-    m_int_flat = _load_field(sidecar, "outputs", "m_int")
-    n_int = k_int_flat.shape[0]
-    print(f"[sphere-pec-driver] Loaded sidecar: fixture_id={sidecar.get('fixture_id', 'N/A')}")
-    print(f"[sphere-pec-driver] n_int (interior edges) = {n_int}")
-
-    # --- Load mesh metadata ---
-    n_nodes = int(sidecar["outputs"]["n_nodes"]["data"][0]) if "n_nodes" in sidecar["outputs"] else None
-    n_tets  = int(sidecar["outputs"]["n_tets"]["data"][0]) if "n_tets" in sidecar["outputs"] else None
-    n_edges = int(sidecar["outputs"]["n_edges"]["data"][0]) if "n_edges" in sidecar["outputs"] else None
-
-    if n_nodes: print(f"[sphere-pec-driver] n_nodes={n_nodes}, n_tets={n_tets}, n_edges={n_edges}")
-
-    # --- Spurious-mode count ---
-    # The predicted spurious count = number of strictly-interior nodes =
-    # spurious_dim in the numpy reference. The TF-Java assembly does not
-    # currently compute this (it would need a second per-node PEC mask pass).
-    # We read it from the NumPy baseline if available, else the caller must
-    # pass --n-spurious.
-    #
-    # From the NumPy baseline: spurious_dim = 368 for the 774-node fixture.
-    # This is safe to hard-code here as a fallback because the fixture is
-    # pinned (sphere.msh is the canonical file). In CI the baseline.json is
-    # always available via the checkout.
-    FIXTURE_SPURIOUS_DIM_FALLBACK = 368
-
-    n_spurious = args.n_spurious
-    if n_spurious is None:
-        # Try to read from the baseline.
-        baseline_path = args.baseline
-        if baseline_path is None:
-            # Try default location relative to sidecar.
-            default_bl = sidecar_path.parent.parent.parent / "fixtures" / "sphere_pec" / "baseline.json"
-            if default_bl.exists():
-                baseline_path = str(default_bl)
-
-        if baseline_path is not None and Path(baseline_path).exists():
-            with open(baseline_path) as f:
-                bl = json.load(f)
-            n_spurious = int(bl["outputs"]["spurious_dim"]["data"][0])
-            print(f"[sphere-pec-driver] spurious_dim from baseline.json = {n_spurious}")
-        else:
-            n_spurious = FIXTURE_SPURIOUS_DIM_FALLBACK
-            print(
-                f"[sphere-pec-driver] WARNING: no baseline.json found; "
-                f"using hard-coded spurious_dim fallback = {n_spurious}"
-            )
-
-    n_request = n_spurious + args.n_request_extra
-    print(f"[sphere-pec-driver] Requesting {n_request} eigenvalues "
-          f"({n_spurious} spurious + {args.n_request_extra} extra)")
-
-    # --- Sanity readouts ---
-    print(f"[sphere-pec-driver] trace(K_int) = {np.trace(k_int_flat):.12e}")
-    print(f"[sphere-pec-driver] trace(M_int) = {np.trace(m_int_flat):.12e}")
-
-    # --- Shift-and-invert ARPACK eigensolve ---
-    k_sp = sp.csr_matrix(k_int_flat)
-    m_sp = sp.csr_matrix(m_int_flat)
-    print("[sphere-pec-driver] Running scipy.sparse.linalg.eigsh (shift-invert, sigma=0)...")
-    eigvals, eigvecs = spla.eigsh(k_sp, k=n_request, M=m_sp, sigma=0.0, which="LM")
-    order = np.argsort(eigvals)
-    eigvals = eigvals[order]
-
-    print(f"[sphere-pec-driver] Recovered {len(eigvals)} eigenvalues.")
-    print(f"[sphere-pec-driver] Lowest {min(5, len(eigvals))} eigenvalues (raw):")
-    for i in range(min(5, len(eigvals))):
-        print(f"  λ[{i}] = {eigvals[i]:.6e}")
-
-    # --- Spurious filter ---
-    if n_spurious + args.k > len(eigvals):
-        print(
-            f"[sphere-pec-driver] ERROR: requested {args.k} physical modes but only "
-            f"{len(eigvals) - n_spurious} available (n_request={n_request}, "
-            f"n_spurious={n_spurious}).",
-            file=sys.stderr,
-        )
-        sys.exit(4)
-
-    physical = eigvals[n_spurious : n_spurious + args.k]
-
-    # Diagnostic ratio (spurious→physical transition).
-    if n_spurious >= 1 and n_spurious < len(eigvals):
-        a = abs(eigvals[n_spurious - 1])
-        b = abs(eigvals[n_spurious])
-        ratio = (b / a) if a > 0.0 else float("inf")
-    else:
-        ratio = float("nan")
-
-    print(f"[sphere-pec-driver] n_spurious = {n_spurious}")
-    print(f"[sphere-pec-driver] spurious→physical ratio = {ratio:.3e}")
-    print(f"[sphere-pec-driver] Lowest {args.k} physical eigenvalues (λ = k²):")
-    for i, lam in enumerate(physical):
-        print(f"  physical[{i}]: λ = {lam:.8e}, k = {np.sqrt(max(lam, 0)):.6f}")
-
-    # --- Cross-backend comparison (if baseline provided) ---
-    if args.baseline and Path(args.baseline).exists():
-        with open(args.baseline) as f:
-            bl = json.load(f)
-        ref_physical = np.array(bl["outputs"]["physical_eigenvalues"]["data"])
-        n_compare = min(len(physical), len(ref_physical))
-        print("\n[sphere-pec-driver] Cross-backend comparison vs NumPy baseline:")
-        print(f"  {'i':>3}  {'TF-Java':>14}  {'NumPy':>14}  {'rel':>10}")
-        all_ok = True
-        for i in range(n_compare):
-            got  = physical[i]
-            want = ref_physical[i]
-            rel  = abs(got - want) / max(abs(want), 1.0)
-            flag = "" if rel < args.rtol else "  <-- FAIL"
-            if rel >= args.rtol:
-                all_ok = False
-            print(f"  {i:>3}  {got:>14.8e}  {want:>14.8e}  {rel:>10.2e}{flag}")
-        if all_ok:
-            print(f"[sphere-pec-driver] PASS: all physical eigenvalues agree to < {args.rtol:.0e} relative.")
-        else:
-            print(f"[sphere-pec-driver] FAIL: some physical eigenvalues exceed {args.rtol:.0e} relative.")
-            sys.exit(5)
-
-    # --- Emit schema-v1 fixture ---
-    n_index = float(sidecar["inputs"]["n_index"]["data"][0]) \
-        if "n_index" in sidecar.get("inputs", {}) else 1.5
-    r_buffer = float(sidecar["inputs"]["r_buffer"]["data"][0]) \
-        if "r_buffer" in sidecar.get("inputs", {}) else 2.0
-
-    result_fixture = {
-        "schema_version": "1",
-        "fixture_id": "sphere_pec/n774_pec_eigenmode_tfjava_eigenresult",
-        "description": (
-            "Physical eigenvalues from the TF-Java sphere-PEC Nédélec assembly + "
-            "SciPy shift-and-invert eigensolve (Epic #88 / #134). "
-            "Cross-checked against reference/fixtures/sphere_pec/baseline.json."
-        ),
-        "units": "lambda = k^2 (inverse-length squared); dimensionless mesh coordinates",
-        "inputs": {
-            "n_index": {
-                "shape": [1], "dtype": "f64",
-                "description": "Refractive index inside the dielectric sphere.",
-                "data": [n_index],
-            },
-            "r_buffer": {
-                "shape": [1], "dtype": "f64",
-                "description": "Outer PEC wall radius.",
-                "data": [r_buffer],
-            },
-            "n_int": {
-                "shape": [1], "dtype": "i64",
-                "description": "Interior edge count (DOFs after PEC elimination).",
-                "data": [n_int],
-            },
-            "n_spurious": {
-                "shape": [1], "dtype": "i64",
-                "description": "Predicted spurious-mode count (interior nodes).",
-                "data": [n_spurious],
-            },
-        },
-        "outputs": {
-            "physical_eigenvalues": {
-                "shape": [args.k], "dtype": "f64",
-                "description": (
-                    f"Lowest {args.k} physical eigenvalues after spurious filtering "
-                    f"(n_spurious={n_spurious}). Acceptance criterion: 1e-5 relative "
-                    "vs baseline.json (Epic #88 cross-IR f64 floor)."
-                ),
-                "tolerance_abs": 1.0e-5,
-                "data": physical.tolist(),
-            },
-            "best_gap": {
-                "shape": [1], "dtype": "f64",
-                "description": (
-                    "Diagnostic ratio lambda[n_spurious] / lambda[n_spurious-1] "
-                    "(spurious-to-physical transition). Large value confirms clean separation."
-                ),
-                "tolerance_abs": 1.0,
-                "data": [ratio],
-            },
-            "eigenvalues_lowest": {
-                "shape": [len(eigvals)], "dtype": "f64",
-                "description": (
-                    f"Raw lowest {len(eigvals)} eigenvalues from shift-and-invert ARPACK "
-                    f"(spurious + physical, before filtering)."
-                ),
-                "tolerance_abs": 1.0e-5,
-                "data": eigvals.tolist(),
-            },
-        },
-        "provenance": {
-            "source": (
-                "reference/tf_java/sphere_pec (assembly) → "
-                "reference/driver/eigensolve_sphere_pec_sidecar.py (SciPy eigensolve seam)"
-            ),
-            "verified_against": "reference/fixtures/sphere_pec/baseline.json",
-            "issue": "#134",
-        },
-    }
-
-    out_path = Path(args.out)
-    out_path.parent.mkdir(parents=True, exist_ok=True)
-    with open(out_path, "w") as f:
-        json.dump(result_fixture, f, indent=2)
-        f.write("\n")
-    print(f"\n[sphere-pec-driver] Wrote {out_path}")
-
-
-if __name__ == "__main__":
-    main()
+sys.argv = [str(Path(__file__).parent / "eigensolve_from_sidecar.py")] + forwarded + extra
+runpy.run_path(sys.argv[0], run_name="__main__")
